Fix player movement mechanics
Add player ground traction and friction, but no more airborne control
Add block 'block_blood'
Explore biome blending

+++ engine/h/core.h
@@ -189,7 +189,7 @@ typedef struct Render
     v2f64 mouse_position;
     v2f64 mouse_last;
     v2f64 mouse_delta;
-    f64 time;
+    u64 time;
     f64 frame_delta;
 } Render;
